+++ pkg/instrumentors/bpf/database/sql/bpf/probe.bpf.c
@@ -1,7 +1,22 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #include "arguments.h"
 #include "span_context.h"
 #include "go_context.h"
 #include "go_types.h"
+#include "uprobe.h"
 
 char __license[] SEC("license") = "Dual MIT/GPL";
 
@@ -9,11 +24,8 @@ char __license[] SEC("license") = "Dual MIT/GPL";
 #define MAX_CONCURRENT 50
 
 struct sql_request_t {
-    u64 start_time;
-    u64 end_time;
+    BASE_SPAN_PROPERTIES
     char query[MAX_QUERY_SIZE];
-    struct span_context sc;
-    struct span_context psc;
 };
 
 struct {
@@ -74,22 +86,4 @@ int uprobe_queryDC(struct pt_regs *ctx) {
 
 // This instrumentation attaches uprobe to the following function:
 // func (db *DB) queryDC(ctx, txctx context.Context, dc *driverConn, releaseConn func(error), query string, args []any)
-SEC("uprobe/queryDC")
-int uprobe_queryDC_Returns(struct pt_regs *ctx) {
-    u64 context_ptr_pos = 3;
-    // Find the corresponding sql event we return from
-    void *context_ptr = get_argument(ctx, context_ptr_pos);
-    void *key = get_consistent_key(ctx, context_ptr);
-    void *sqlReq_ptr = bpf_map_lookup_elem(&sql_events, &key);
-
-    struct sql_request_t sqlReq = {0};
-    bpf_probe_read(&sqlReq, sizeof(sqlReq), sqlReq_ptr);
-    sqlReq.end_time = bpf_ktime_get_ns();
-
-    // Send event
-    bpf_perf_event_output(ctx, &events, BPF_F_CURRENT_CPU, &sqlReq, sizeof(sqlReq));
-    // Clean the sql event
-    bpf_map_delete_elem(&sql_events, &key);
-    stop_tracking_span(&sqlReq.sc);
-    return 0;
-}
+UPROBE_RETURN(queryDC, struct sql_request_t, 3, 0, sql_events, events)